feat(#678): move treasury aside detail pages
