--- apps/web/src/app/[lang]/dho/[id]/@aside/(.)treasury/deposit/loading.tsx
@@ -1,10 +1,0 @@
-import { DepositFunds } from '@hypha-platform/epics';
-import { SidePanel } from '../../_components/side-panel';
-
-export default async function Treasury() {
-  return (
-    <SidePanel>
-      <DepositFunds closeUrl={''} spaceId={0} isLoading={true} />
-    </SidePanel>
-  );
-}
