+++ apps/web/src/app/[lang]/dho/[id]/@aside/treasury/deposit/loading.tsx
@@ -1,0 +1,9 @@
+import { DepositFunds, SidePanel } from '@hypha-platform/epics';
+
+export default async function Treasury() {
+  return (
+    <SidePanel>
+      <DepositFunds closeUrl={''} spaceId={0} isLoading />
+    </SidePanel>
+  );
+}
