--- apps/web/src/app/[lang]/dho/[id]/treasury/constants.ts
@@ -1,5 +1,0 @@
-import { Locale } from '@hypha-platform/i18n';
-
-export const getDhoPathTreasury = (lang: Locale, id: string) => {
-  return `/${lang}/dho/${id}/treasury`;
-};
